Make UserService backend runtime configurable: Fixes #1445
