+++ horreum-backend/src/main/java/io/hyperfoil/tools/horreum/svc/user/DatabaseUserBackend.java
@@ -1,4 +1,4 @@
-package io.hyperfoil.tools.horreum.svc;
+package io.hyperfoil.tools.horreum.svc.user;
 
 import io.hyperfoil.tools.horreum.api.internal.services.UserService;
 import io.hyperfoil.tools.horreum.entity.user.Team;
@@ -7,12 +7,10 @@ import io.hyperfoil.tools.horreum.entity.user.TeamRole;
 import io.hyperfoil.tools.horreum.entity.user.UserInfo;
 import io.hyperfoil.tools.horreum.entity.user.UserRole;
 import io.hyperfoil.tools.horreum.server.WithRoles;
-import io.quarkus.arc.Unremovable;
-import io.quarkus.arc.properties.IfBuildProperty;
-import io.quarkus.security.Authenticated;
+import io.hyperfoil.tools.horreum.svc.Roles;
+import io.hyperfoil.tools.horreum.svc.ServiceException;
+import io.quarkus.arc.lookup.LookupIfProperty;
 import io.quarkus.security.identity.SecurityIdentity;
-import jakarta.annotation.security.PermitAll;
-import jakarta.annotation.security.RolesAllowed;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.persistence.criteria.CriteriaBuilder;
@@ -27,46 +25,39 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
 
+import static io.hyperfoil.tools.horreum.svc.UserServiceImpl.getTeamPrefix;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
 
-@PermitAll
 @ApplicationScoped
-@Unremovable
-@IfBuildProperty(name = "horreum.roles.provider", stringValue = "database")
-public class DatabaseUserServiceImpl implements UserService {
+@LookupIfProperty(name = "horreum.roles.provider", stringValue = "database")
+public class DatabaseUserBackend implements UserBackEnd {
 
-   private static final Logger LOG = Logger.getLogger(DatabaseUserServiceImpl.class);
+   @Inject
+   SecurityIdentity identity;
 
-   @Inject SecurityIdentity identity;
+   private static final Logger LOG = Logger.getLogger(DatabaseUserBackend.class);
 
-   private static UserData toUserInfo(UserInfo info) {
-      return new UserData("", info.username, info.fistName, info.lastName, info.email);
+   private static UserService.UserData toUserInfo(UserInfo info) {
+      return new UserService.UserData("", info.username, info.fistName, info.lastName, info.email);
    }
 
-   @Authenticated
-   @Override public List<String> getRoles() {
-      return identity.getRoles().stream().toList();
-   }
-
-   @Authenticated
    @WithRoles(extras = Roles.HORREUM_SYSTEM)
-   @Override public List<UserData> searchUsers(String query) {
+   @Override public List<UserService.UserData> searchUsers(String query) {
       List<UserInfo> users = UserInfo.list("username like ?1", "%" + query + "%");
-      return users.stream().map(DatabaseUserServiceImpl::toUserInfo).collect(toList());
+      return users.stream().map(DatabaseUserBackend::toUserInfo).collect(toList());
    }
 
    @WithRoles(extras = Roles.HORREUM_SYSTEM)
-   @RolesAllowed({Roles.VIEWER, Roles.TESTER, Roles.ADMIN})
-   @Override public List<UserData> info(List<String> usernames) {
+   @Override public List<UserService.UserData> info(List<String> usernames) {
       List<UserInfo> users = UserInfo.list("username in ?1", usernames);
-      return users.stream().map(DatabaseUserServiceImpl::toUserInfo).collect(toList());
+      return users.stream().map(DatabaseUserBackend::toUserInfo).collect(toList());
    }
 
    @Transactional
    @WithRoles(fromParams = NewUserParameterConverter.class)
-   @RolesAllowed({Roles.MANAGER, Roles.ADMIN})
-   @Override public void createUser(NewUser user) {
+   @Override
+   public void createUser(UserService.NewUser user) {
       if (user == null) {
          throw ServiceException.badRequest("Missing user as the request body");
       } else if (user.team != null && !user.team.endsWith("-team")) {
@@ -104,43 +95,22 @@ public class DatabaseUserServiceImpl implements UserService {
       userInfo.persist();
    }
 
+   public static final class NewUserParameterConverter implements Function<Object[], String[]> {
+      @Override public String[] apply(Object[] objects) {
+         return new String[] { ((UserService.NewUser) objects[0]).user.username };
+      }
+   }
    private void addTeamMembership(UserInfo userInfo, String teamName, TeamRole role) {
       Optional<Team> storedTeam = Team.find("teamName", teamName).firstResultOptional();
       userInfo.teams.add(new TeamMembership(userInfo, storedTeam.orElseGet(() -> Team.getEntityManager().merge(new Team(teamName))), role));
    }
-
-   public static final class NewUserParameterConverter implements Function<Object[], String[]> {
-      @Override public String[] apply(Object[] objects) {
-         return new String[] { ((NewUser) objects[0]).user.username };
-      }
-   }
-
-   @Authenticated
    @Override public List<String> getTeams() {
       List<Team> teams = Team.listAll();
       return teams.stream().map(t -> t.teamName + "-team").collect(toList());
    }
-
-   @WithRoles(addUsername = true)
-   @Override public String defaultTeam() {
-      UserInfo userInfo = UserInfo.findById(identity.getPrincipal().getName());
-      return userInfo != null ? userInfo.defaultTeam : null;
-   }
-
-   @Transactional
-   @WithRoles(addUsername = true)
-   @Override public void setDefaultTeam(String team) {
-      UserInfo userInfo = UserInfo.findById(identity.getPrincipal().getName());
-      userInfo.defaultTeam = Util.destringify(team);
-      userInfo.persistAndFlush();
-   }
-
-   @Authenticated
    @WithRoles(extras = Roles.HORREUM_SYSTEM)
-   @Override public Map<String, List<String>> teamMembers(String team) {
-      if (!identity.getRoles().contains(getTeamPrefix(team) + Roles.MANAGER) && !identity.getRoles().contains(Roles.ADMIN)) {
-         throw ServiceException.badRequest("This user is not a manager for team " + team);
-      }
+   @Override
+   public Map<String, List<String>> teamMembers(String team) {
       Team teamEntity = Team.find("teamName", team.substring(0, team.length() - 5)).firstResult();
       if (teamEntity == null) {
          throw ServiceException.notFound("The team " + team + " does not exist");
@@ -152,12 +122,8 @@ public class DatabaseUserServiceImpl implements UserService {
    }
 
    @Transactional
-   @Authenticated
    @WithRoles(fromParams = UpdateTeamMembersParameterConverter.class)
    @Override public void updateTeamMembers(String team, Map<String, List<String>> roles) {
-      if (!identity.getRoles().contains(getTeamPrefix(team) + Roles.MANAGER) && !identity.getRoles().contains(Roles.ADMIN)) {
-         throw ServiceException.forbidden("This user is does not have the manager role for team " + team);
-      }
       Team teamEntity = Team.find("teamName", team.substring(0, team.length() - 5)).firstResult();
       if (teamEntity != null) {
          roles.forEach((username, teamRoles) -> {
@@ -174,28 +140,26 @@ public class DatabaseUserServiceImpl implements UserService {
          throw ServiceException.notFound("The team " + team + " does not exist");
       }
    }
-   
+
    public static final class UpdateTeamMembersParameterConverter implements Function<Object[], String[]> {
       @SuppressWarnings("unchecked") @Override public String[] apply(Object[] objects) {
          return ((Map<String, List<String>>) objects[1]).keySet().toArray(new String[0]);
       }
    }
 
-   @RolesAllowed(Roles.ADMIN)
+   @WithRoles(extras = Roles.HORREUM_SYSTEM)
    @Override public List<String> getAllTeams() {
       List<Team> teams = Team.listAll();
       return teams.stream().map(t -> t.teamName + "-team").collect(toList());
    }
-
    @Transactional
-   @RolesAllowed(Roles.ADMIN)
+   @WithRoles(extras = Roles.HORREUM_SYSTEM)
    @Override public void addTeam(String team) {
       Team.getEntityManager().merge(new Team(team.substring(0, team.length() - 5)));
    }
 
    @Transactional
-   @RolesAllowed(Roles.ADMIN)
-   @WithRoles
+   @WithRoles(extras = Roles.HORREUM_SYSTEM)
    @Override public void deleteTeam(String team) {
       String prefix = getTeamPrefix(team);
       Team teamEntity = Team.find("teamName", prefix.substring(0, prefix.length() - 1)).firstResult();
@@ -207,27 +171,12 @@ public class DatabaseUserServiceImpl implements UserService {
       }
    }
 
-   private String getTeamPrefix(String team) {
-      if (team == null || team.isBlank()) {
-         throw ServiceException.badRequest("No team name!");
-      } else if (team.startsWith("horreum.")) {
-         throw ServiceException.badRequest("Team name starting with 'horreum.' is illegal; this is reserved for internal use.");
-      } else if (!team.endsWith("-team")) {
-         throw ServiceException.badRequest("Team name must end with '-team' suffix");
-      } else if (team.length() > 64) {
-         throw ServiceException.badRequest("C'mon, can you think on a shorter team name?");
-      }
-      return team.substring(0, team.length() - 4);
-   }
-
-   @RolesAllowed(Roles.ADMIN)
    @WithRoles(extras = Roles.HORREUM_SYSTEM)
-   @Override public List<UserData> administrators() {
-      return getAdministratorUsers().stream().map(DatabaseUserServiceImpl::toUserInfo).collect(toList());
+   @Override public List<UserService.UserData> administrators() {
+      return getAdministratorUsers().stream().map(DatabaseUserBackend::toUserInfo).collect(toList());
    }
 
    @Transactional
-   @RolesAllowed(Roles.ADMIN)
    @WithRoles(extras = Roles.HORREUM_SYSTEM)
    @Override public void updateAdministrators(List<String> newAdmins) {
       if (!newAdmins.contains(identity.getPrincipal().getName())) {
