+++ horreum-backend/src/main/resources/application.properties
@@ -107,7 +107,7 @@ horreum.db.secret=secret
 ## horreum.keycloak.url=http://localhost:8180
 horreum.keycloak.realm=horreum
 horreum.keycloak.clientId=horreum-ui
-horreum.roles.provider=database
+horreum.roles.provider=keycloak
 
 # Address used in links to Horreum
 horreum.url=http://localhost:3000
